% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asSparseMatrix}
\alias{asSparseMatrix}
\title{Create sparse matrix from data.frame}
\usage{
asSparseMatrix(df, checkNA = TRUE)
}
\arguments{
\item{df}{: data.frame}

\item{checkNA}{: set false if you can gualantee there is no NAs in the df.}
}
\description{
Create sparse matrix from data.frame
}

