% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{thread_function2}
\alias{thread_function2}
\title{apply funciton to data.frame column}
\usage{
thread_function2(df, n)
}
\arguments{
\item{df}{: data.frame}

\item{n}{: number of cores for parallel operation}
}
\description{
apply funciton to data.frame column
}

